Collapsed shard_client into cas_client
