--- shard_client/Cargo.toml
@@ -1,53 +1,0 @@
-[package]
-name = "shard_client"
-version = "0.14.5"
-edition = "2021"
-
-[features]
-strict = []
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-utils = {path = "../utils"}
-file_utils = {path = "../file_utils" }
-merklehash = {path = "../merklehash"}
-cas_client = {path = "../cas_client"}
-xet_error = {path = "../xet_error"}
-mdb_shard = {path = "../mdb_shard"}
-merkledb = {path = "../merkledb"}
-cas_types = {path = "../cas_types"}
-tokio = { version = "1.36", features = ["full"] }
-tokio-retry = "0.3.0"
-tower = {version = "0.4"}
-clap = "2.33"
-async-trait = "0.1.9"
-anyhow = "1"
-http = "0.2.5"
-serde_json = "1.0"
-tempfile = "3"
-tracing = "0.1.31"
-bincode = "1.3.3"
-uuid = {version = "1", features = ["v4", "fast-rng"]}
-lazy_static = "1.4.0"
-heed = "0.11"
-itertools = "0.10"
-# use embedded webpki root certs for MacOS as native certs take a very long time
-# to load, which affects startup time significantly
-[target.'cfg(macos)'.dependencies]
-reqwest = { version = "0.11.4", features = ["json", "webpki-roots"] }
-[target.'cfg(not(macos))'.dependencies]
-reqwest = { version = "0.12.7", features = ["json"] }
-reqwest-middleware = "0.3.3"
-url = "2.5.2"
-bytes = "1"
-
-# trace-propagation
-opentelemetry = { version = "0.17", features = ["trace", "rt-tokio"] }
-opentelemetry-jaeger = { version = "0.16", features = ["rt-tokio"] }
-opentelemetry-http = "0.6.0"
-tracing-opentelemetry = "0.17.2"
-
-[dev-dependencies]
-rand = "0.8.5"
-
